function_apps/requirements.txt
xmltodict

# Azure SDKs
azure-identity
azure-mgmt-resource 
azure-mgmt-web